+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### next release
-* Add Varanus::SSL::CSR.generate
+* Added Varanus::SSL::CSR.generate
+* Added Reports
+  * Varanus::Reports#ssl - list of SSL/TLS certs
+  * Varanus::Reports#domains - list of domains validated with DCV
 
 ### 0.1.0 (2018-11-07)
 * Initial release